+++ stocks/market/src/main/java/nl/rug/aoop/market/transaction/Transaction.java
@@ -9,7 +9,7 @@ import nl.rug.aoop.market.stock.Stock;
  * @param stockPrice The price at which the stock was traded.
  * @param stock The stock that was traded.
  */
-public record Transaction(@Getter Stock stock, @Getter int stockPrice, @Getter int amount){
+public record Transaction(@Getter Stock stock, @Getter double stockPrice, @Getter long amount){
     /**
      * Custom constructor for the Transaction that updates the stock price accordingly.
      * @param stock The stock that was traded.
